ceva

+++ src/Person.java
@@ -2,7 +2,7 @@
  * Created by George Stratulat on 03/03/2017.
  */
 public class Person {
-    private double age;
+    private int age;
     private String name;
 
     public int getAge() {
